refactor: mv runStake dir to stakeMint
