+++ packages/inter-protocol/exported.js
@@ -1,3 +1,3 @@
-import './src/runStake/types.js';
+import './src/stakeMint/types.js';
 import './src/vaultFactory/types.js';
 import './src/vpool-xyk-amm/types.js';
