+++ packages/inter-protocol/src/stakeMint/attestation.js
