+++ packages/inter-protocol/test/attestation/test-attestation.js
@@ -7,7 +7,7 @@ import { E, Far } from '@endo/far';
 
 import { makeCopyBag } from '@agoric/store';
 import { setupZCFTest } from '@agoric/zoe/test/unitTests/zcf/setupZcfTest.js';
-import { makeAttestationFacets } from '../../src/runStake/attestation.js';
+import { makeAttestationFacets } from '../../src/stakeMint/attestation.js';
 
 const { details: X } = assert;
